mviznARMG/requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
constantly==15.1.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@151c15a0e5c81b9edebec1c5111f97e1098c9b0a
entrypoints==0.3
gast==0.3.3
google-pasta==0.1.8
GPUtil==1.4.0
grpcio==1.27.2
h5py==2.10.0
hyperlink==19.0.0
idna==2.8
imageio==2.6.1
imgaug==0.3.0
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.18.1
pandas==1.1.2
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.6.7
ptyprocess==0.6.0
pycocotools==2.0.2
Pygments==2.5.2
PyHamcrest==2.0.0
pylibmc==1.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
scikit-image==0.16.2
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
SharedArray==3.2.0
six==1.14.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-serving-api==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
torch==1.6.0
torchvision==0.7.0
traitlets==4.3.3
Twisted==19.10.0
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.0
zipp==2.2.0
zope.interface==4.7.1
